Compressor state prediction with LSTM and Kalman filter
The data were denoised and normalized with the Kalman filter.
Using K-Means, compressor status labels were determined intelligently.
LSTM model was designed in PyTorch and trained to predict compressor condition.
The accuracy of the model was checked with Confusion Matrix and it had an acceptable performance.
The model was saved in ONNX format to be ready for use in API and Flask server.